Move model helpers out of utils, move PaymentsByOrderIdLoader to payments
