+++ saleor/graphql/order/types.py
@@ -21,7 +21,8 @@ from ...order import OrderStatus, models
 from ...order.models import FulfillmentStatus
 from ...order.utils import get_order_country, get_valid_shipping_methods_for_order
 from ...payment import ChargeStatus
-from ...payment.utils import get_total_captured
+from ...payment.dataloaders import PaymentsByOrderIdLoader
+from ...payment.model_helpers import get_total_captured
 from ...product.product_images import get_product_image_thumbnail
 from ..account.dataloaders import AddressByIdLoader, UserByUserIdLoader
 from ..account.types import User
@@ -55,7 +56,6 @@ from .dataloaders import (
     OrderEventsByOrderIdLoader,
     OrderLineByIdLoader,
     OrderLinesByOrderIdLoader,
-    PaymentsByOrderIdLoader,
 )
 from .enums import OrderEventsEmailsEnum, OrderEventsEnum
 from .utils import validate_draft_order
